+++ src/root.zig
@@ -3,7 +3,9 @@ const std = @import("std");
 pub const db = @import("db.zig");
 pub const http = @import("http.zig");
 pub const migrations = @import("migrations.zig");
+pub const password = @import("password.zig");
 pub const server = @import("server.zig");
+pub const users = @import("users.zig");
 
 test {
     std.testing.refAllDecls(@This());